% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_save.R
\name{bray_save}
\alias{bray_save}
\title{Save png output}
\usage{
bray_save(path, plot, w = 5, h = 5, dpi = 300)
}
\arguments{
\item{path}{Destination for png output}

\item{plot}{ggplot2 plot object}

\item{w}{width in inches (Default = \code{5})}

\item{h}{height in inches (Default = \code{5})}

\item{dpi}{density per inch (Default = \code{300})}
}
\value{
Nothing
}
\description{
Save png output
}
