% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddplot.R
\name{ddplot}
\alias{ddplot}
\title{Generalization of QQplot with depth measures}
\usage{
ddplot(x, y, colorx = "#e8e7e3", colory, method = "Tukey")
}
\arguments{
\item{x}{First dataframe}

\item{y}{Second dataframe}

\item{colorx}{HTML code for color of points from first dataframe}

\item{colory}{HTML code for color of points from second dataframe}

\item{method}{One of the available depth functions (Default = \code{"Tukey"})}
}
\value{
ggplot
}
\description{
Generalization of QQplot with depth measures
}
\examples{
ddplot(df.1, df.2, colory = "#ff0000")
}
