% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bray.R
\name{theme_bray}
\alias{theme_bray}
\title{Theme object for ggplot2}
\usage{
theme_bray(
  maincolor = "#e8e7e3",
  offcolor = "#66696e",
  gridh = TRUE,
  gridv = FALSE
)
}
\arguments{
\item{maincolor}{HTML formatted color string for default objects like points, lines, ...(Default = \code{"#e8e7e3"})}

\item{offcolor}{HTML formatted color string for secondary objects like gridlines, axes, ...(Default = \code{"#66696e"})}

\item{gridh}{Bool flag for major horizontal gridlines (Default = \code{TRUE})}

\item{gridv}{Bool flag for major vertical gridlines (Default = \code{FALSE})}
}
\value{
Theme object
}
\description{
Theme object for ggplot2
}
\examples{
theme_bray(gridh = FALSE)
}
